fix: buttons switch lang move to  count verse
